Merge rewrite into main

+++ Melodify/src/main/java/bg/sofia/uni/fmi/melodify/controller/AlbumController.java
@@ -3,32 +3,50 @@ package bg.sofia.uni.fmi.melodify.controller;
 import bg.sofia.uni.fmi.melodify.dto.AlbumDto;
 import bg.sofia.uni.fmi.melodify.mapper.AlbumMapper;
 import bg.sofia.uni.fmi.melodify.model.Album;
+import bg.sofia.uni.fmi.melodify.service.AlbumCreateFacadeService;
 import bg.sofia.uni.fmi.melodify.service.AlbumService;
+import bg.sofia.uni.fmi.melodify.service.AlbumSetFacadeService;
+import bg.sofia.uni.fmi.melodify.service.TokenManagerService;
 import bg.sofia.uni.fmi.melodify.validation.ApiBadRequest;
+import bg.sofia.uni.fmi.melodify.validation.MethodNotAllowed;
 import bg.sofia.uni.fmi.melodify.validation.ResourceNotFoundException;
+import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
+
+import static bg.sofia.uni.fmi.melodify.security.RequestManager.isAdminByRequest;
 
 @RestController
 @RequestMapping(path = "api/albums")
 @Validated
 public class AlbumController {
     private final AlbumService albumService;
+    private final AlbumCreateFacadeService albumCreateFacadeService;
+    private final AlbumSetFacadeService albumSetFacadeService;
+    private final TokenManagerService tokenManagerService;
     private final AlbumMapper albumMapper;
 
-    public AlbumController(AlbumService albumService, AlbumMapper albumMapper) {
+    @Autowired
+    public AlbumController(AlbumService albumService, AlbumCreateFacadeService albumCreateFacadeService,
+                           AlbumSetFacadeService albumSetFacadeService, TokenManagerService tokenManagerService,
+                           AlbumMapper albumMapper) {
         this.albumService = albumService;
+        this.albumCreateFacadeService = albumCreateFacadeService;
+        this.albumSetFacadeService = albumSetFacadeService;
+        this.tokenManagerService = tokenManagerService;
         this.albumMapper = albumMapper;
     }
 
     @GetMapping
-    public List<AlbumDto> getAllAlbums() {
-        return albumMapper.toDtoCollection(albumService.getAllAlbums());
+    public List<AlbumDto> getAlbums(@RequestParam Map<String, String> filters) {
+        return albumMapper.toDtoCollection(albumService.getAlbums(filters));
     }
 
     @GetMapping(value = "/{id}")
@@ -47,8 +65,21 @@ public class AlbumController {
     @PostMapping
     public Long addAlbum(@NotNull(message = "The provided album description in the body cannot be null")
                          @RequestBody
-                         AlbumDto albumDto) {
-        Album potentialAlbumToCreate = albumService.createAlbum(albumMapper.toEntity(albumDto));
+                         AlbumDto albumDto,
+                         @RequestParam("genre_id")
+                         @NotNull(message = "The provided genre id cannot be null")
+                         @Positive(message = "The provided genre id must be positive")
+                         Long genreId,
+                         @RequestParam("artist_ids")
+                         @NotNull(message = "The provided artist ids cannot be null")
+                         List<Long> artistIdsList,
+                         HttpServletRequest request) {
+        if (!isAdminByRequest(request, tokenManagerService)) {
+            throw new MethodNotAllowed("There was a problem in authorization");
+        }
+
+        Album potentialAlbumToCreate = albumCreateFacadeService
+            .createAlbumWithGenreAndArtists(albumMapper.toEntity(albumDto), genreId, artistIdsList);
 
         if (potentialAlbumToCreate == null) {
             throw new ApiBadRequest("There was a problem in creating the album");
@@ -61,7 +92,12 @@ public class AlbumController {
     public AlbumDto deleteAlbumById(@RequestParam("album_id")
                                     @NotNull(message = "The provided album id cannot be null")
                                     @Positive(message = "The provided album id must be positive")
-                                    Long albumId) {
+                                    Long albumId,
+                                    HttpServletRequest request) {
+        if (!isAdminByRequest(request, tokenManagerService)) {
+            throw new MethodNotAllowed("There was a problem in authorization");
+        }
+
         return albumMapper.toDto(albumService.deleteAlbum(albumId));
     }
 
@@ -72,7 +108,17 @@ public class AlbumController {
                                 Long id,
                                 @RequestBody
                                 @NotNull(message = "The provided album dto in the body cannot be null")
-                                AlbumDto albumToUpdate) {
-        return albumService.setAlbumById(albumToUpdate, id);
+                                AlbumDto albumToUpdate,
+                                @RequestParam(name = "genre_id", required = false)
+                                Long genreId,
+                                @RequestParam(name = "artist_ids", required = false)
+                                List<Long> artistIdsList,
+                                HttpServletRequest request) {
+        if (!isAdminByRequest(request, tokenManagerService)) {
+            throw new MethodNotAllowed("There was a problem in authorization");
+        }
+
+        return albumSetFacadeService.setAlbumWithGenreAndArtistsIfProvided(id, albumMapper.toEntity(albumToUpdate),
+            genreId, artistIdsList);
     }
 }
